Update imports after renaming modules to libs
