+++ api/ErrorReports/indexes.js
@@ -1,4 +1,4 @@
-import createIndex from '../../../modules/server/createIndex';
+import createIndex from '../../../libs/server/createIndex';
 import ErrorReports from '../ErrorReports';
 
 createIndex(ErrorReports, { createdById: 1 });
