+++ api/ErrorReports/schemas/error-report.js
@@ -4,7 +4,7 @@ import SimpleSchema from 'simpl-schema';
 import i18n from 'meteor/universe:i18n';
 /* #### PLOP_IMPORTS_START #### */
 /* #### PLOP_IMPORTS_END #### */
-import parseObjectID from '../../../modules/parseObjectID';
+import parseObjectID from '../../../libs/parseObjectID';
 
 const cleanLevel = (item) => parseInt(item, 10);
