+++ api/Roles/server/indexes.js
@@ -1,4 +1,4 @@
-import createIndex from '../../../modules/server/createIndex';
+import createIndex from '../../../libs/server/createIndex';
 import Roles from '../Roles';
 
 // this index is added by the alanning:roles module automatically
