+++ tools/plop/generators/AddEmailTemplate/templates/moduleObject/action.js.hbs
@@ -4,7 +4,7 @@ import { Meteor } from 'meteor/meteor';
 // import { gql } from '@apollo/client';
 // import { getGraphqlSchema } from '{{pathOffset apiFolderName}}../../../startup/server/graphql-api';
 import {{pascalCase apiFolderName }} from '../{{pascalCase apiFolderName }}';
-import sendEmail from '{{pathOffset apiFolderName}}../../../modules/server/sendEmail';
+import sendEmail from '{{pathOffset apiFolderName}}../../../libs/server/sendEmail';
 
 const send{{pascalCase name }}Email = ({ userId, {{camelCase apiFolderName }}Id }, context) => {
   const {{ singular (camelCase apiFolderName) }} = {{pascalCase apiFolderName }}.findOne({ {{camelCase apiFolderName }}Id });
