+++ tools/plop/generators/AddEmailTemplate/templates/notifyAdmins/action.js.hbs
@@ -4,7 +4,7 @@ import { Meteor } from 'meteor/meteor';
 // import { gql } from '@apollo/client';
 // import { getGraphqlSchema } from '{{pathOffset apiFolderName}}../../../startup/server/graphql-api';
 // import {{pascalCase apiFolderName }} from '../{{pascalCase apiFolderName }}';
-import sendEmail from '{{pathOffset apiFolderName}}../../../modules/server/sendEmail';
+import sendEmail from '{{pathOffset apiFolderName}}../../../libs/server/sendEmail';
 
 const send{{pascalCase name }}Email = (data, context) => {
   Meteor.users.find({ roles: 'admin' }).forEach((user) => {
