+++ tools/plop/generators/BasicApiModule/templates/api/mutations.js.hbs
@@ -2,7 +2,7 @@ import {{ pascalCase pluralName }} from './{{ pascalCase pluralName }}';
 import {{ pascalCase singularName }}Schema from './schemas/{{ dashCase singularName }}';
 // import sanitizeHtml from 'sanitize-html';
 import checkUserRole from '{{apiPathOffset}}../Users/actions/checkUserRole';
-import createMongoModifier from '{{apiPathOffset}}../../modules/server/createMongoModifier';
+import createMongoModifier from '{{apiPathOffset}}../../libs/server/createMongoModifier';
 
 export default {
   save{{ pascalCase singularName }}: (root, args, { user }) => {
