+++ tools/plop/generators/DocumentModule/templates/api/server/indexes.js.hbs
@@ -1,4 +1,4 @@
-import createIndex from '{{apiPathOffset}}../../../modules/server/createIndex';
+import createIndex from '{{apiPathOffset}}../../../libs/server/createIndex';
 import {{ pascalCase pluralName }} from '{{apiPathOffset}}../{{ pascalCase pluralName }}';
 
 createIndex({{ pascalCase pluralName }}, { owner: 1 });
