+++ tools/plop/generators/DocumentModule/templates/ui/View{{ pascalCase singularName }}.jsx.hbs
@@ -8,7 +8,7 @@ import BlankState from '../components/BlankState';
 import FileIcon from '@ant-design/icons/FileOutlined';
 import Comments from './components/Comments';
 import { {{ camelCase singularName }} as {{ camelCase singularName }}Query } from './queries/{{ pascalCase pluralName }}.gql';
-import parseMarkdown from '../../modules/parseMarkdown';
+import parseMarkdown from '../../libs/parseMarkdown';
 import CommentComposer from './components/CommentComposer';
 
 import { StyledView{{ pascalCase singularName }}, {{ pascalCase singularName }}Body } from './StyledView{{ pascalCase singularName }}';
