+++ tools/plop/generators/DocumentModule/templates/ui/components/Comments.jsx.hbs
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import CommentComposer from './CommentComposer';
-import { timeago } from '../../../modules/dates';
+import { timeago } from '../../../libs/dates';
 
 import { StyledComments, CommentsList, CommentsListHeader, Comment } from './StyledComments';
