+++ tools/plop/generators/DocumentModule/templates/ui/{{ pascalCase pluralName }}.jsx.hbs
@@ -4,7 +4,7 @@ import Button from 'antd/lib/button';
 import { Link } from 'react-router-dom';
 import { useQuery, useMutation } from '@apollo/client';
 import i18n from 'meteor/universe:i18n';
-import { timeago } from '../../modules/dates';
+import { timeago } from '../../libs/dates';
 import BlankState from '../components/BlankState';
 import FileIcon from '@ant-design/icons/FileOutlined';
 import Loading from '../components/Loading';
