+++ tools/plop/generators/StubModule/templates/api/mutations.js.hbs
@@ -2,7 +2,7 @@
 // import {{ pascalCase singularName }}Schema from './schemas/{{ dashCase singularName }}';
 // import sanitizeHtml from 'sanitize-html';
 // import checkUserRole from '{{apiPathOffset}}../Users/actions/checkUserRole';
-// import createMongoModifier from '{{apiPathOffset}}../../modules/server/createMongoModifier';
+// import createMongoModifier from '{{apiPathOffset}}../../libs/server/createMongoModifier';
 
 export default {
   /* #### PLOP_MUTATIONS_START #### */
