+++ tools/plop/generators/StubModule/templates/api/schemas/{{ dashCase singularName }}.js.hbs
@@ -160,7 +160,7 @@ import i18n from 'meteor/universe:i18n';
 {{/each}}
 /* #### PLOP_IMPORTS_END #### */
 {{#compare fieldTypes 'includes' 'ObjectID'}}
-import parseObjectID from '{{apiPathOffset}}../../../modules/parseObjectID';
+import parseObjectID from '{{apiPathOffset}}../../../libs/parseObjectID';
 {{/compare}}
 
 {{#each schema.fields }}
