+++ tools/plop/generators/StubModule/templates/ui/{{ pascalCase pluralName }}.jsx.hbs
@@ -11,7 +11,7 @@ import Row from 'antd/lib/row';
 import Col from 'antd/lib/col';
 import PageBreadcrumbs, { Breadcrumb } from '{{uiPathOffset}}../components/PageBreadcrumbs';
 import PageHeader from '{{uiPathOffset}}../components/PageHeader';
-// import hasRole from '{{uiPathOffset}}../../modules/hasRole';
+// import hasRole from '{{uiPathOffset}}../../libs/hasRole';
 
 import Styled{{ pascalCase pluralName }} from './Styled{{ pascalCase pluralName }}';
