+++ tools/plop/templates/filter-default.js.hbs
@@ -1,7 +1,7 @@
 {{#compare template '===' 'ui/imports'}}
 import isString from 'lodash/isString';
 {{#if reference}}
-import useQueryForFilters from '{{@root.uiPathOffset}}../../../modules/hooks/useQueryForFilters';
+import useQueryForFilters from '{{@root.uiPathOffset}}../../../libs/hooks/useQueryForFilters';
 {{/if}}
 {{/compare}}
 {{#compare template '===' 'ui/helpers'}}
