+++ ui/components/Validation.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import validate from '../../modules/validate';
+import validate from '../../libs/validate';
 
 class Validation extends React.Component {
   constructor(props) {
