+++ ui/error-reports/components/ErrorReportsTable.jsx
@@ -6,8 +6,8 @@ import Table from 'antd/lib/table';
 import isString from 'lodash/isString';
 import i18n from 'meteor/universe:i18n';
 import { useHistory } from 'react-router-dom';
-import hasRole from '../../../modules/hasRole';
-import useQueryStringObject from '../../../modules/hooks/useQueryStringObject';
+import hasRole from '../../../libs/hasRole';
+import useQueryStringObject from '../../../libs/hooks/useQueryStringObject';
 import PrettyDate from '../../components/PrettyDate';
 import NewErrorReportButton from './NewErrorReportButton';
