+++ ui/roles/EditRole.jsx
@@ -4,7 +4,7 @@ import i18n from 'meteor/universe:i18n';
 import { useQuery /* , useMutation */ } from '@apollo/client';
 import Divider from 'antd/lib/divider';
 import { useParams } from 'react-router-dom';
-import hasRole from '../../modules/hasRole';
+import hasRole from '../../libs/hasRole';
 import Loading from '../components/Loading';
 import PageBreadcrumbs, { Breadcrumb } from '../components/PageBreadcrumbs';
 import PageHeader from '../components/PageHeader';
