+++ ui/users/admin/components/AdminUsersTable.jsx
@@ -5,7 +5,7 @@ import queryString from 'query-string';
 // import { Roles } from 'meteor/alanning:roles';
 import i18n from 'meteor/universe:i18n';
 import { useQuery } from '@apollo/client';
-import useQueryStringObject from '../../../../modules/hooks/useQueryStringObject';
+import useQueryStringObject from '../../../../libs/hooks/useQueryStringObject';
 import isString from 'lodash/isString';
 import Table from 'antd/lib/table';
